Rearrange code.
