+++ internal/patch/patch.go
@@ -6,6 +6,7 @@ import (
 	f "github.com/elimity-com/scim/internal/filter"
 	"github.com/elimity-com/scim/schema"
 	"github.com/scim2/filter-parser/v2"
+	"strings"
 )
 
 // Op represents the possible value the operation is to perform.
@@ -84,4 +85,64 @@ func (v OperationValidator) Validate() (interface{}, error) {
 	default:
 		return nil, fmt.Errorf("invalid operation op: %s", v.op)
 	}
+}
+
+// getRefAttribute returns the corresponding attribute based on the given attribute path.
+//
+// e.g.
+//  - `userName` would return the `userName` attribute.
+//	- `name.givenName` would return the `givenName` attribute.
+//  - `ext:employeeNumber` would return the `employeeNumber` attribute from the extension.
+func (v OperationValidator) getRefAttribute(attrPath filter.AttributePath) (*schema.CoreAttribute, error) {
+	// Get the corresponding schema, this can be the main schema or an extension.
+	var refSchema = v.schema
+	if uri := attrPath.URI(); uri != "" {
+		// It can also be an extension if it has a uri prefix.
+		var ok bool
+		if refSchema, ok = v.schemas[uri]; !ok {
+			return nil, fmt.Errorf("invalid uri prefix: %s", uri)
+		}
+	}
+
+	// Get the correct attribute corresponding to the given attribute path.
+	var (
+		refAttr  *schema.CoreAttribute
+		attrName = attrPath.AttributeName
+	)
+	for _, attr := range refSchema.Attributes {
+		if strings.EqualFold(attr.Name(), attrName) {
+			refAttr = &attr
+			break
+		}
+	}
+	if refAttr == nil {
+		return nil, fmt.Errorf("could not find attribute %s", v.path)
+	}
+	if subAttrName := attrPath.SubAttributeName(); subAttrName != "" {
+		refSubAttr, err := v.getRefSubAttribute(refAttr, subAttrName)
+		if err != nil {
+			return nil, err
+		}
+		refAttr = refSubAttr
+	}
+	return refAttr, nil
+}
+
+// getRefSubAttribute returns the sub-attribute of the reference attribute that matches the given subAttrName, if none
+// are found it will return an error.
+func (v OperationValidator) getRefSubAttribute(refAttr *schema.CoreAttribute, subAttrName string) (*schema.CoreAttribute, error) {
+	if !refAttr.HasSubAttributes() {
+		return nil, fmt.Errorf("the referred attribute has no sub-attributes: %s", v.path)
+	}
+	var refSubAttr *schema.CoreAttribute
+	for _, attr := range refAttr.SubAttributes() {
+		if strings.EqualFold(attr.Name(), subAttrName) {
+			refSubAttr = &attr
+			break
+		}
+	}
+	if refSubAttr == nil {
+		return nil, fmt.Errorf("could not find attribute %s", v.path)
+	}
+	return refSubAttr, nil
 }
